IGP-134: Update USDai ecosystem launch limits and deprecate vault 174
- Action 1: Raise USDai ecosystem limits to launch-scale for DEXes 46-48 and vaults 171-173, 175-179, and 180, while removing Team Multisig auth on all markets except DEX 47 and vault 180.
- Action 2: Deprecate wrongly deployed vault 174 (USDai/USDC) with a full pause and remove its Team Multisig auth.
- Action 3: Claim accrued iETHv2 (Lite) stETH revenue to Team Multisig.

Also, update simulation setup to deploy vault 180 if not already live on the fork.
